requirements.txt
python-docx>=0.8.11
transformers>=4.40.0
torch>=2.0.0
sentence-transformers>=2.6.0
accelerate>=0.29.0
API_KEY = "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"